Add files via upload
This .txt file contains a direct link to the YouTube channel whose video analytics data is central to the analysis in this project. The link provides access to the channel, allowing for easy reference to the source of the videos analyzed in the repository. This enables viewers or collaborators to explore the channel’s content and observe firsthand the videos associated with the engagement metrics and performance trends examined in the project. This link enhances transparency by connecting the raw data and findings in the repository to their real-world source.